documentation-tools: doxygen automatic dependency generation
Some primitive doxygen depency generation added. Based on the doxygen
configuration file we try to guess the .h.c dependencies. However,
doxygen can have directories + patterns in which case we are living
on our wild guesses how doxygen finds the dependencies that are
not in the same directory.

MRP_DOCINIT autoconf macro was also introduced to initiate the
documentation building