common: minor transport changes, transport test fixes.
Renamed transport events and requests, the generic messaging versions
recv* and send* are now called recvmsg* and sendmsg*. Changed accept
to not take event callbacks for the new accepted transport, instead
the callbacks need to be given to the listening transport and are
inherited from there.

Added basic event callback checks: connection-oriented transports
require connection, close, and recvmsg*. Connectionless transports
require recvfrom*. For connect (usually sender-filtering) to succeed
on a connectionless transports, it needs to have a recv* callback
specified.