debug: semi-automatic infra to resolve debug file:line to function name
Unfortunately starting with 4.x proper support for __FUNCTION__ has
been degenerated to an alias to the C99-compatible func. Since func
is an automatic variable instead of an automatic macro, many of the
interesting stuff (such as constant-string concatenation) you cannot
do with it that were possible with __FUNCTION__.

This is a slight setback for our debug infra because the intention
was to set up the main debugging macro, mrp_debug, in such a way that
it would have autogenerated for every debug site a string table entry
of the format __DEBUG_SITE_function@file:line. Listing the existing
sites would have then simply become finding all matching entries in the
string table and printing them out.

Now with the degenerated >= 4.x gcc, we only get file and line number
information but not function name. To overcome this, support has been
added to the debug infra for registering the necessary information to
do the extra file:line to function name mapping. Moreover, a simple
script and some common boilerplate code has been added to help make
use of all this without having to really understand the details of how
it all works, or having to touch any of your code (other than placing
mrp_debug's wherever appropriate).

You can take a look at src/Makefile.am and check how the common or core
libraries or the dbus plugin is built to see what you need to do to
automatically generate the necessary debug function info.