WiP-core: murphy lua bindings prototype for basic daemon configuration.
This prototype provides equivalent functionality to the built-in
configuration language of the daemon. IOW, it provides functions
for

  - testing the existence of plugins
  - loading plugin instances with optional parameters
  - basic info, error, warning message logging

This proto currently does not use or depend on the common lua-utils
or lua-decision 'convenience libraries'. However it can be easily
moved on top of those if needed. This would probably make sense if
we start adding bindings to more complex bits of functionality
(such as the mainloop, transports, events, etc).