lua-bindings: added support for evaluating Lua from the console.
You can now evaluate Lua code snippets or Lua files from the console.
Please note that currently you need to fully quote Lua code you enter
via the console to protect it from the console input tokenizer. This
limitation will be lifted once we have implemented support for non-
pre-tokenized console commands.

Also note that all output produced by Lua code is currently going to
the standard output of the murphy daemon and not to the console where
you entered the Lua snippet. Consider this a bug which should be fixed.
Fixing this is a bit harder that it was with the old ohm-time console
implementation. In all its crudeness (whenever 'grabbed') the old ohm
console was simply dup2ing stdout to the socket fd of the console
connection during the execution of console commands. With our current
console things this is not so easy: we have mrp_console_printf to print
to the console which is not fd-based but is using glibc's custom stream
handler support...