resource-dbus: added support for configuring the D-Bus bus.
This is supposed to be used for zone support. For instance, if there
are multiple session buses in the system, they can be assigned to
different zones by running several resource-dbus plugin instances.  At
the moment there is only one-to-one mapping between zones and buses,
but this could change later.