mainloop-test: omit glib pump if we have a glib superloop.
Don't try to run with glib set up as both our super- and subloops.
It makes no sense and it only ends up in glib's event dispatcher
being called recursively (superloop dispatching us, we dispatching
the subloop which happens to be also our superloop) and aborting
on an assert.

This whole single combined mainloop-test mess should be cleaned up.