daemon: delayed log target activation
To quickly enable the activation of loggers registered by plugins
the log target is now activated from the daemon startup main code
path (used to be done during command line parsing).

This is a quick hack to get dlog going. Probably the right thing
to do is to try activating it as early as possible (during command
line parsing) to get startup messages going to the right place,
take a note if this fails and in this case retry it from the daemon
startup main code path. Later...