websockets: added support for pure HTTP websockets and file serving.
This patch adds support for pure HTTP sockets. Because connection
establishment is tied to websocket handshaking in libwebsockets, a
considerable amount of extra gymnastics is needed to get pure HTTP
clients, which don't handshake, be symmetric to genuine websocket
clients which do handshake.

We need to emulate the connection establishment event ourselves on
demand when the first HTTP content request comes in from a pure
HTTP client. Additionally we also need to handle the mapping from
libwebsockets ws's to our wsl_sck_t's differently because for pure
HTTP clients libwebsockets refuses to allocate user data even if
the registered user data size is non-zero for protocol 0. Pfoof...