websockets: configurable limited HTTP content serving over transports.
It is now possible to serve content over HTTP on a websocket transport.
This is primarily intended to allow handing out javascript libraries that
interface the actual service exposed over the same websocket transport.
The javascript library can hide the dirty details of the underlying
communication protocol and present a more conventional and user-friendly
javascript API. By serving the interface library on demand from the same
transport as the service itself, the protocol can be freely reworked
behind the scenes, ideally keeping high-level API intact, without having
to worry about compatibility issues between the javascript library and
the protocol being used. In a way the library is an agent pushed by the
protocol implementation to the client...