common: added wakeup (sort of passive deferred) callback support.
Added support for wakeup callbacks to the murphy mainloop. A wakeup
callback is not quite unlike a deferred one. The crucial difference
is that wakeup callbacks do not prevent the mainloop from blocking
on (e)poll. wakeup callbacks are not executed if the mainloop is
dispatching because of active deferred callbacks (approximated by
checking if poll_timeout is 0). Using wakeup callbacks one can piggy-
back some processing on wakeups that were trigered by I/O (or signal)
events or expiring timers.