common: make sure *alloc(0) returns NULL.
The C standard allows [mc]alloc(0) to return a unique non-NULL pointer
instead of NULL and indeed the stock allocator in recent versions of
glibc does behave this way. We prefer our zero-sized chunks of memory
to be represented by NULL, so we take care of this in the allocator
wrappers.