domain-control: javascript API hack + a small demo/test
Hacked together a javacript API and wrote a small test that imports
the audio and video playback owners from murphy and exports a single
increasing integer counter to murphy. Well, how should I put this...
the javascript API needs more work. This is more of a PoC proto than
the real thing. Especially when it comes to error handling.