daemon: valgrind re-exec support, source tree detection
The daemon can now re-exec itself through valgrind, preserving its
command line and the environment if so requested.

Also the daemon tries to detect if it is run directly from the
source tree. If it is, the runtime defaults are modified to
load the configuration file and plugins from the same source
tree and to run with increased verbosity. If necessary these
can be overridden on the command line.