common: more versatile data serialization via native-types.
Added native-types (and a primitive TLV encoder/decoder) to
provide more flexible and versatile data/type serialization.

Ideally (and eventually) we'd like to be able to throw any
natively declared data type at any supported transport and
have that automatically take care of the serialization and
deserialization necessary for transmission. Although with
this initial implementation we're far not there yet, this
should already provide a much more flexible framework to
work with and later build upon than what the original
MRP_TRANSPORT_MODE_DATA is capable of providing.

There are still quite a few rough edges and features either
missing or not fully implemented, including but not limited
to

  - support for unions,
  - implement blobs,
  - automatic handling of linked lists of known types
  - support all the (sensible) type/layout combinations
  - allow self- and cross-referencing data types (IOW
    allow forward-references to yet undeclared types, and
    provide a function for checking if a type is complete,
    ie. there are no undeclared types it references, and
    this holds true also for all its referenced types)

Eventually we'd like to get to the point where, instead of
having to keep the limitations of the available transport
encoding and decoding in mind, one could freely design data
types as seen fit, and the transports could always handle
them.