lua-utils: support for typed objects, arrays, and autobridged members.
Extended funcbridge to support type-checked arrays and (murphy
core/lua-utils/object-infra) objects. The extensions include

  - signature extensions for object and array specifications:
      o [s|d|f]:     an array of strings, integers, or doubles
      o [*]:         an array of any of the above types
      o O(<C-type>): a Lua object of the given registered Murphy C type
      o O(*):        a Lua object of any of the registered Murphy C types
  - collecting, type-checking, and passing arrays to bridged calls
  - fetching, type-checking, and passing objects to bridged calls

Additionally, as calls to funcbridge member methods (from Lua to C)
arrive at funcbridge with an extra table at the bottom of the stack,
this needs special treatment if we want to make the bridged calls
with a genuine-looking stack. Funcbridge now has the necessary code
to patch the stack before the bridge (taking care of preventing any
garbage collection of temporarily removed objects), if the bridged
callee has indicated its desire for this.