common: allow setting a mapping table of native-type id's for transports.
Added support for a native-type id mapping table for transports. This
is necessary to avoid the problems arising from not having the exactly
same set of native-types registered (or having the same set registered
in different order) by the communicating parties. With this mechanism,
it is enough for the communicating parties to agree about the ids of
the native-types passed over that particular instance of transport
instead of having to agree about the ids of all the registered native-
types over all the transports.

The support is still a bit lacky on the user-friendliness side. We'll
probably need a few additional macros and convenience functions to
allow people to easily set up a mapping table correctly.