common: make sure the logger is never recursing
Make sure the logger is never recursively invoked. The default logger
backend can never recurse. However with an overridden backend this
could happen, especially if the backend pushes log messages over a
transport or some other generic IPC. If something goes wrong at the
IPC level, an error is usually logged. If this results in an attempt
to use the same IPC we might end up recursing ad infinitum, or rather
ad stack-overflowtum...

Since we can't safely assume backends to properly check and handle
such a corner case, it is better to protect against this in the
common infra.