lua-utils: try harder getting complex object getter/setter cases right.
Try to handle better cases when a (potentially extensible) object
class has both predefined members and an overridden setter/getter
with or without native/whitelisted members. The current semantics
is roughly the following:

  1) check and handle if the field corresponds to a predefined member
  2) if there is a setfield/getfield declared and the field is marked
     native or no fields have been marked native at all, pass it to
     setfield/getfield
  3) if there is no setfield/getfield or there is one but it did not
     handle the field (returned 0), pass it to setext/getext if the
     class is extensible

Note that in principle we should treat it an error if a field declared
as native is not handled by setfield/getfield. Currently we don't.

Also note that with these semantics it probably would be better to
forget about marking members as native and just remove native
whitelisting altogether. We'd then instead always pass every non-
predeclared field to setfield/getfield, then try setext/getext if
this setfield/getfield did not handle the field (IOW returned 0).

This would also get us rid of the potential ambiguity of what to do
if a field has been declared native but then setfield/getfield does
not handle it.