lua-utils: patch the stack for {s,g}etfield during init + fix a bad pattern of thinkos.
This could/should still be improved a bit. Now setfield and getfield
must use top-relative addressing and also be extremely careful not
to screw up the stack by popping it too much. Probably it would be
much betters to patch the stack and then invoke setfield or getfield
via lua_pcall which would both protect and clean up the stack for us.
This'll probably need another iteration...

This patch also fixes a pattern of horribly embarrassing thinkos in
the setter/getter and initializer status checking / fallback selection
logic.