lua-bindings: override allocator earlier, use simpler block tracking.
Looks like in all hastiness I missed that it is indeed possible
to override the Lua allocator before it has been used to allocate
any blocks. So... instead of blaming the Lua infra for a non-
existing deficiency, let's try to use it correctly. Instead of
first creating the interpreter state and only then overriding the
allocator, let's use the state creation routine which also takes
the allocator to use as an argument.

In all fairness, it is IMO still a deficiency that the interpreter
lets a busy allocator (one that has unfreed blocks) be overridden
without storing and passing the used allocators userdata along
with the allocated blocks of memory.

Now correctly overriding the allocator with a tracking one at _any_
time has a pretty steep performance penalty. The allocator first
needs to check if a block has been allocated by it and otherwise
pass it on to the previously active one. This is what we used to do
before this patch. With this patch, we choose lower runtime overhead
for the price of less flexibility (can't enable tracking dynamically
at any time).