lua-utils: allow object tracking to be enabled from the outside.
Added a function to enable/disable object tracking externally.
Currently this is used (in lua-murphy) to automatically turn on
object tracking whenever Lua memory tracking is enabled. Note
that once the first object has been allocated, the state of
tracking cannot be altered since it would have an effect on the
size of object userdata.