resource: experimental patch for unlimited number of resources.
This is a proof-of-concept patch to remove the inherent limitation
of the supported number of resources. The patch takes the most
straightforward approach possible without any optimizations by
replacing the original uint32_t mask with a dynamically sized
mask array.

Note that the patch provides only the first step in lifting the
limitations. The hardcoded statically sized data structures that
depend on the original *MAX_RESOURCES macros are still statically
sized with *MAX_RESOURCES hardcoded to 1024 for testing. This
needs further patching to get rid of those without losing too
much performance.

Also note that the Javascript library of the wrt-resource plugin
has not been updated to correctly handle the protocol changes.
That too is left for another day...