common: first shot at improving native-types.
This is my first shot at rolling a more flexible version of
native-types. Introduced an alternative native type registry
that supports more versatile data types, serialization and
deserialization. In addition to the basic data types supported
by native-types, this one also has support for

    - handling of cross-referencing data types
    - flexible arrays
    - unions (although only limited support)

There are at least a few things we need to still add, including
but not limited to

    - printing/dumping instances of types
    - support for custom encoders/decoders (for D-Bus)
    - maybe a way of specifying resource-limits for decoding

Plus a cleanup...