common: mainloop support for edge-triggered I/O events.
This patch adds support for specifying on a per-watch basis whether
I/O events should be level or edge-triggered and also for setting
and retrieving the default I/O event trigger mode.

Because we provide both default and per-watch settings, we need an
extra flag for indicating level-triggered mode for a watch when the
default is set to EPOLLET (MRP_IO_TRIGGER_EDGE). This makes the patch
slightly more complex than just providing the necessary enum flags.