libresource: Info instead of error when ignoring events
The printed message is used for both informational purposes
as well as when there is an error state - and with error
states there are always other error messages printed, so
having this message be printed out by default is not
necessary. Modifies the printed message, as ignoring events
is not specific to cases of malformed data.

As a side effect, this stops "ignoring malformed resource
event" error messages from coming up in the log when
receiving the first message from Murphy (could be replicated
with both resource-api-test as well as my Python wrapper).