common: event infrastructure overhaul, lifted from core to common.
This patch litfs the original event publish/subscribe code from
core to common and attempts to fix some of its shortcomings.

The patch also introduces events busses to provide an abstraction
for logically grouping related or similar events together. Arbitrary
number of busses can be created dynamically. Subscription and event
delivery (emitting) is always done on a particular bus. There is
a default global (the NULL) bus which is always availablei.
This is the easiest to use for backward-compatibility for code which
used to generate or subscribe to some events but did not have direct
access to the mainloop.

Events can now be delivered both synchronously and asynchronously.
The event mechanism have been put under the mainloop mainly for this
reason (to automatically associate busses with a mainloop) to support
asynchronous delivery. With synchronous delivery, the event have been
delivered to all subscribers by the time the emitting call returns.
With asynchronous delivery the event will be delivered from a deferred
callback once the mainloop is done processing its pending events. The
global NULL bus is not associated with a mainloop and therefore only
supports synchronous delivery.

Data can now be attached to events in different formats. In addition
to the original mrp_msg_t, we now support JSON (mrp_json_t) and custom
pointers. For mrp_msg_t and JSON, the infrastructure takes care of
properly reference-counting the data in case of asychronous delivery.
For custom data this is not done ATM (although that might change), so
the caller has to make sure the event data is valid during delivery
and properly freed afterwards.

Currently it is not possible to delete event busses or unregister
events.