lua-event: adjusted to event infra changes, added explicit start/stop.
Updated to use the revamped event infrastructure. Added explicit
methods for starting/stopping watches. Added oneshot member to be
more in line with timers and deferred callbacks. Minor stylistical
changes to be more in line with the rest of the Lua bindings.

Emitting events from Lua with data attached is not implemented yet...