common: first shot at an object extension mechanims.
The Murphy object extension mechanism aims to provide a lightweight
easy to use mechanism for extending objects at runtime with arbitrary
data. The implementation tries to strike the right balance between
being flexible enough, with low runtime overhead, while also providing
limited support for runtime type-checking.

The primary intention behind the object extension mechanism is to help
us keeping the Murphy infrastructure generic and reusable. We can keep
the common infra clean, free of any domain-specific data and let
dynamically loaded plugins use this mechanism to attach application-
and domain-specific data to the common infra as needed. A typical example
could be, for instance, writing a plugin that provides Python bindings
for the Murphy mainloop. If runtime extensions are enabled for the
mainloop data structure, the plugin can directly hook its own necessary
data for the Python runtime to the mainloop as an extension.