common: wrote a much cleaner, more efficent hash-table implementation.
Reworked the hash-table implementation. It is pretty much a rewrite
from scratch to provide a much cleaner implementation, more consistent
API, and to address some of the complains about the original one.

The biggest change, however, is the support for O(1) hash table entry
access (for insert, delete, lookup, and replace operations) via the
introduction of table entry cookies. The usage of cookies is completely
optional and the hash table is fully functional without cookies. Proper
usage of cookies, however, can significantly speed up access to table
entries, practically making the hash table almost a fast as an array.

Cookies can be thought of as virtual- or pseudo-indices of hash table
entries. They are not used as direct indices, rather a bit more like
virtual addresses in an OS. They can be mapped to the proper table
entry very fast, using a few instructions.

The implementation supports both user- and table-generated cookies. The
ones generated by the table itself will make sure that access to entries
is properly accelerated. However, it is easy to generate proper cookies
with similar properties if/when necessary. The easiest way is to use
cookies, unique per-entry integers, that are between 1 and the number of
entries in the table.

Another significant change is how tables can be iterated. The much
criticised/hated callback-based iterating foreach function is gone. An
iterating FOREACH macro is provided instead, on top of which callback-
based iterating functions similar to the original one can be trivially
implemented.

Also an API has been added to limit memory usage of both any particular
hash table instance and all hash table instances collectively. Although
the API is mostly in place, it is pretty limited at the moment, and the
logic to check and limit memory usage for actual overallocation is still
not in place.