core: added pointer for GLib main context and loop.
Added context pointers for GMainLoop and GMainContext in case
someone wants/needs to set up the mainloop to be pumped by a
GMainLoop. The next patch adds the necessary changes to the
daemon itself.

I hope this to be a temporary workaround and that we I'll be
able to get rid of this once the current problems with threaded
GMainLoop subloops are solved.