websocklib: fix broken poll handling for some libwebsockets versions.
Rolled a quick fix to address two separate issues people have been
running into on Ubuntu.

The smaller and easier of these is just to allow compilation with
libwebsockets-1.2.2-1 (in the Ubuntu scheme of versioning). Looks
like this version is new enough to have context creation info but
too old to support setting/overriding cipher lists. If someone
tries to set the cipher list for such a build now (via setopt),
we produce an error message and fail.

The bigger and somewhat more difficult to catch problem was that
although this version already has the new polling interface, it
still uses the old scheme of passing fd (user, vs. in). This ended
poll being set up for fd 0 (in happened to be NULL) insted of the
real HTTP socket being used which practically rendered the websocket
interface inactive. Even worse, hitting enter (IOW stdin/fd 0) while
running the daemon in the foreground drove the daemon into busyish
100 % CPU-consuming loop (not to a real busyloop though, so you
couldn't easily catch this unless enough debugging was turned on).

This patch tries to detect the pollfd passing conventions, which
can be one of the following:

  - user: the oldest implementation, fd in <user>
  - in: used before struct pollargs was introduced, fd in <in>
  - pollargs: fd and event mask in a struct pointed to by <in>

Unfortunately, libwebsockets usually does not flip version numbers
or define macros when an incompatible change is introduces which
makes it sometimes difficult to compile-time detect which version
we are compiling against. For this change, we try to approximate
the difference between user and in passing conventions by checking
for the presence of LWS_MAX_HEADER_NAME_LENGTH. This macro was
removed 3 days after the switch from <user> to <in> with only a
3 or so patches in between. Because this still leaves a commit time-
window of a few patches that are misdetected, we compensate for this
runtime by checking if on the <in> path it *looks* like fd is passed
still via <user> and fix things up if this happens to be the case.
This is not a fully reliable test but the way we use libwebsockets
this seems to be good enough...

BTW, it looks like if one thing we can rely on, it is our bad luck:
the libwebsockets version in Ubuntu really seems to fall into this
misdetection window of 3 or so commits, so the runtime fixup logic
is needed. How about that, folks...